fix(live-assist): backend runtime — ESM boot split, STT heartbeat, transcript publish, DB-backed devices
- move buildOrchestrator to server/api/liveAssistBoot.ts with ESM imports (inline require() crashed the backend at boot)
- markSttAlive(): treat the STT config poll as a liveness heartbeat so the status bar reflects 'service running', not just 'segments flowing'
- publishTranscript(): emit each finalized segment on the live-assist channel
- SettingsService STT devices persisted in DB so the backend's report is visible to the Next.js process (separate processes)
- GET suggestions returns sttStatus for panel init
