+++ server/backend.ts
@@ -34,7 +34,8 @@ import obsRouter from "./api/obs";
 import twitchRouter from "./api/twitch";
 import mediaPlayerRouter from "./api/media-player";
 import wordHarvestRouter from "./api/word-harvest";
-import { createLiveAssistRouter, buildOrchestrator } from "./api/live-assist";
+import { createLiveAssistRouter } from "./api/live-assist";
+import { buildOrchestrator } from "./api/liveAssistBoot";
 import { APP_PORT, BACKEND_PORT, WS_PORT } from "../lib/config/urls";
 import { createServerWithFallback } from "../lib/utils/CertificateManager";
 import { getPortConflictReport } from "../scripts/port-diagnostics.mjs";
